+++ crates/atuin-daemon/src/search/index.rs
@@ -7,6 +7,7 @@
 //! - Frecency-based ranking (frequency + recency)
 //! - Dynamic filtering by directory, host, session, etc.
 
+use super::normalize_diacritics;
 use std::{borrow::Cow, collections::HashSet, sync::Arc};
 
 use atuin_client::history::{History, is_known_agent};
@@ -227,37 +228,6 @@ pub enum IndexFilterMode {
 /// the last rebuild sit past the end and score 0 until the next rebuild.
 type FrecencyMap = Arc<Vec<u32>>;
 
-/// Longest query the fuzzy matcher will see. Frizbee's `u16` scores overflow
-/// (and panic) somewhere past ~2700 needle chars; no real query is anywhere
-/// near either limit, so longer input is truncated.
-const MAX_QUERY_LEN: usize = 512;
-
-/// Truncate a query to the longest length frizbee can score without
-/// panicking in `Matcher::from_query`. Anything that hands a query to
-/// frizbee (including client-side highlighting) must apply this.
-pub fn truncate_query(query: &str) -> &str {
-    // O(1) happy path -- query cannot exceed `MAX_QUERY_LEN` chars if it doesn't even have that
-    // many bytes.
-    if query.len() <= MAX_QUERY_LEN {
-        return query;
-    }
-    match query.char_indices().nth(MAX_QUERY_LEN) {
-        Some((end, _)) => &query[..end],
-        None => query,
-    }
-}
-
-/// Normalize Latin diacritics to their ASCII equivalents (`é` → `e`) so
-/// unaccented queries match accented history entries. Maps char to char so
-/// char positions are preserved.
-pub fn normalize_diacritics(s: &str) -> Cow<'_, str> {
-    use super::normalize::normalize;
-    if s.is_ascii() || !s.chars().any(|c| normalize(c) != c) {
-        return Cow::Borrowed(s);
-    }
-    Cow::Owned(s.chars().map(normalize).collect())
-}
-
 /// One entry in the fuzzy matcher's haystack: the original string plus the normalized version
 /// (diacritics removed) we actually match against.
 struct HaystackEntry {
@@ -388,7 +358,7 @@ impl SearchIndex {
         // Get precomputed frecency map (may be None if not yet computed)
         let frecency_map = self.frecency_map.read().await.clone();
 
-        let query = truncate_query(query);
+        let query = super::truncate_query(query);
         // Match accent-insensitively: the haystack side is normalized in
         // add_history, so an accented query must be normalized too
         let query = normalize_diacritics(query);
